Binary-Trade-Placer/requirements.txt
python-telegram-bot
requests
python-dotenv
pandas
schedule
